requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
requests==2.31.0
pypdf==3.17.1
python-docx==1.1.0
gunicorn==21.2.0
python-dotenv==1.0.0

# AI (fixed + compatible)
sentence-transformers==2.7.0
transformers==4.41.2
huggingface_hub==0.23.0
numpy==1.26.2

# CPU-only torch (VERY IMPORTANT)
torch==2.2.2+cpu
